adding an administration system with adding, update, delete and publi…

+++ App/templates/single.php
@@ -12,15 +12,20 @@ $this->title = 'Post';
 Utils::echoFlashBag('message');
 ?>
 
-<h1>Mon blog</h1>
+<div class="text-center"><h1>Article</h1></div>
 
 <div>
     <h2><?= htmlspecialchars(ucfirst($post->getTitle()));?></h2>
     <h5><?= htmlspecialchars($post->getChapo());?></h5>
     <p><?= htmlspecialchars($post->getContent());?></p>
     <p><strong><?= htmlspecialchars(ucfirst($post->getAuthor()));?></strong></p>
-    <p>Crée le : <?= htmlspecialchars($post->getDateAdded());?><br>
-        Modifié le : <?= htmlspecialchars($post->getDateAmended()) ?></p>
+    <p>Crée le : <?= htmlspecialchars($post->getDateAdded());?><br><?php
+        if (!empty($post->getDateAmended())) {
+        ?>
+        Modifié le : <?= htmlspecialchars($post->getDateAmended())?></p>
+        <?php
+        }
+        ?>
 </div>
 <div>
     <h4>Commentaires</h4>
@@ -40,10 +45,15 @@ Utils::echoFlashBag('message');
     {
     ?>
         <p><strong><?= htmlspecialchars(ucfirst($comment->getUsername()));?></strong></p>
-        <p><?=htmlspecialchars($comment->getContent());?></p>
-        <p>Posté le : <?=htmlspecialchars($comment->getDateAdded());?><br>
-        Modifié le :<?= htmlspecialchars($comment->getDateAmended())?><br><?php
-        if(($_SESSION['userId'] === $comment->getUserId()) OR $_SESSION['is_admin'] === 1) {?>
+        <p><?=htmlspecialchars(ucfirst($comment->getContent()));?></p>
+        <p>Posté le : <?=htmlspecialchars($comment->getDateAdded());?><br><?php
+            if (!empty($comment->getDateAmended())) {
+            ?>
+            Modifié le : <?= htmlspecialchars($comment->getDateAmended())?></p>
+            <?php
+                    }
+
+        if(isset($_SESSION['userId']) && (( $_SESSION['userId'] === $comment->getUserId()) OR $_SESSION['is_admin'] === 1)) {?>
             <a href="index.php?p=delete-comment&idArt=<?=$comment->getPostId()?>&idComment=<?=$comment->getId()?>">Supprimer</a>
             <a href="index.php?p=update-comment&idArt=<?=$comment->getPostId()?>&idComment=<?=$comment->getId()?>">Modifier</a></p>
             <?php
